SPR-8532 Upgrade org.springframework.web.servlet to Servlet 3.0 (as provided dependency) and add support for javax.servlet.Part parameter
git-svn-id: https://src.springframework.org/svn/spring-framework/trunk@4713 50f2f4bb-b051-0410-bef5-90022cba6387

+++ org.springframework.web.servlet/src/test/java/org/springframework/web/servlet/resource/ResourceHttpRequestHandlerTests.java
@@ -60,10 +60,11 @@ public class ResourceHttpRequestHandlerTests {
 		handler.handleRequest(request, response);
 		assertEquals("text/css", response.getContentType());
 		assertEquals(17, response.getContentLength());
-		assertTrue(((Long)response.getHeader("Expires")) >= System.currentTimeMillis() - 1000 + (3600 * 1000));
+		assertTrue(Long.valueOf(response.getHeader("Expires")) >= System.currentTimeMillis() - 1000 + (3600 * 1000));
 		assertEquals("max-age=3600, must-revalidate", response.getHeader("Cache-Control"));
 		assertTrue(response.containsHeader("Last-Modified"));
-		assertEquals(response.getHeader("Last-Modified"), new ClassPathResource("test/foo.css", getClass()).getFile().lastModified());
+		assertEquals(Long.valueOf(response.getHeader("Last-Modified")).longValue(), 
+				new ClassPathResource("test/foo.css", getClass()).getFile().lastModified());
 		assertEquals("h1 { color:red; }", response.getContentAsString());
 	}
 	
@@ -75,10 +76,11 @@ public class ResourceHttpRequestHandlerTests {
 		MockHttpServletResponse response = new MockHttpServletResponse();
 		handler.handleRequest(request, response);
 		assertEquals("text/html", response.getContentType());
-		assertTrue(((Long)response.getHeader("Expires")) >= System.currentTimeMillis() - 1000 + (3600 * 1000));
+		assertTrue(Long.valueOf(response.getHeader("Expires")) >= System.currentTimeMillis() - 1000 + (3600 * 1000));
 		assertEquals("max-age=3600, must-revalidate", response.getHeader("Cache-Control"));
 		assertTrue(response.containsHeader("Last-Modified"));
-		assertEquals(response.getHeader("Last-Modified"), new ClassPathResource("test/foo.html", getClass()).getFile().lastModified());
+		assertEquals(Long.valueOf(response.getHeader("Last-Modified")).longValue(), 
+				new ClassPathResource("test/foo.html", getClass()).getFile().lastModified());
 	}
 	
 	@Test
@@ -90,10 +92,11 @@ public class ResourceHttpRequestHandlerTests {
 		handler.handleRequest(request, response);
 		assertEquals("text/css", response.getContentType());
 		assertEquals(17, response.getContentLength());
-		assertTrue(((Long)response.getHeader("Expires")) >= System.currentTimeMillis() - 1000 + (3600 * 1000));
+		assertTrue(Long.valueOf(response.getHeader("Expires")) >= System.currentTimeMillis() - 1000 + (3600 * 1000));
 		assertEquals("max-age=3600, must-revalidate", response.getHeader("Cache-Control"));
 		assertTrue(response.containsHeader("Last-Modified"));
-		assertEquals(response.getHeader("Last-Modified"), new ClassPathResource("testalternatepath/baz.css", getClass()).getFile().lastModified());
+		assertEquals(Long.valueOf(response.getHeader("Last-Modified")).longValue(), 
+				new ClassPathResource("testalternatepath/baz.css", getClass()).getFile().lastModified());
 		assertEquals("h1 { color:red; }", response.getContentAsString());
 	}
